+++ test/test313.js
@@ -12,14 +12,20 @@ describe('Test 313 SEARCH ORDER BY', function() {
 
     var data = [{a:1},{a:10},{a:2}];
     var res = alasql('SEARCH / ORDER BY (a) FROM ?', [data]);
+    assert.deepEqual(res,[ { a: 1 }, { a: 2 }, { a: 10 } ]);
 
     var data = [1,10,2];
     var res = alasql('SEARCH / ORDER BY (_) FROM ?', [data]);
+    assert.deepEqual(res,[ 1, 2, 10 ]);
 
     var data = [{a:1},{a:10},{a:2}];
     var res = alasql('SEARCH / ORDER BY (a DESC) a  FROM ?', [data]);
+    assert.deepEqual(res,[ 10, 2, 1 ]);
 
-    console.log(res);
+    var data = [{a:1, b:10},{a:10, b:0},{a:2,b:7}];
+    var res = alasql('SEARCH / ORDER BY (a+b) FROM ?', [data]);
+    assert.deepEqual(res,[ { a: 2, b: 7 }, { a: 10, b: 0 }, { a: 1, b: 10 } ]);
+
     done();
 
   });